requirements.txt
allure-pytest==2.6.2
allure-python-commons==2.6.2
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
enum34==1.1.6
idna==2.8
importlib-metadata==0.18
lxml==4.3.4
more-itertools==7.1.0
namedlist==1.7
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==3.8.2
pytest-cov==2.7.1
pytest-ordering==0.6
pytest-repeat==0.8.0
pytest-rerunfailures==7.0
pytest-sugar==0.9.2
requests==2.22.0
selenium==3.141.0
six==1.12.0
termcolor==1.1.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
